requirements.txt
transformers
torch
fastapi
uvicorn
accelerate
openai
python-dotenv